Less opinionated root path
If a studio didn't want to use a mountpoint or root in their file tree and they set the values to empty strings get_root_path() would still return a string with separators. This change allows a studio to pass empty string for either mountpoint, root, or both and still get a valid path.